New-ADuser
This is what I have working for my network I recently was tasked with new Duties so I had to come up with something to create multiple accounts quickly and provision them and move them to the correct OU with the click of a button. This is what I found and used it and crafted it to my Companies needs.